requirements.txt
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
bandit==1.6.2
decorator==4.4.2
Django==3.0.8
django-extensions==3.0.3
django-filter==2.3.0
djangorestframework==3.11.0
flake8==3.8.3
gitdb==4.0.5
GitPython==3.1.7
importlib-metadata==1.7.0
ipdb==0.13.3
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
Markdown==3.2.2
mccabe==0.6.1
parso==0.7.1
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
smmap==3.0.4
sqlparse==0.3.1
stevedore==3.2.0
traitlets==4.3.3
wcwidth==0.2.5
zipp==3.1.0
